+++ src/main/data/duke.txt
@@ -1,3 +1,3 @@
-[D][ ] do hw (by: Aug 30 2023, 23:59)
+[D][X] do hw (by: Aug 30 2023, 23:59)
 [E][ ] PINUS (from: Aug 30 2023, 15:00 to: Aug 30 2023, 16:00)
 [E][ ] SOMETHING (from: Aug 31 2023, 12:00 to: Aug 31 2023, 13:00)
